+++ internal/gamemanager/discovery_other.go
@@ -1,0 +1,7 @@
+//go:build !windows
+
+package gamemanager
+
+func discoverWindowsDSTRoots() []string {
+	return nil
+}